dont update alias pubkey on xfer if xfer pubkey points to alias
caused a connectblock problem on a random peer who got an alias transfer
but the xfer pubkey was already in an alias (a timing issue?).

So i fail to transfer the pubkey in this case, the alias wont reaelly be
usable by the new person so it kinda goes in limbo, but better than
having a chain that wont sync.


Former-commit-id: 211dbbcc29306dcedc53601ad4de4987ad98559f